v0.1.1
There we go again...
Ok. I was thinking a lot about interfaces and in my opinion it wasn't good idea to deal with them in this type of game(at least until now) because of focus on mouse events - it's not good trying to force interfaces in update method with pressedor entered button.
I will think about it later (maybe it's good option for taskable actions) but there will be no many distinguishable classes so at this point i dont see any purposes to do this in that way.